requirements.txt
langchain
langchain-community
langchain-google-genai
google-generativeai
Streamlit
python-dotenv
requests
beautifulsoup4
duckduckgo-search
google-search-results